+++ CHANGES.md
@@ -11,6 +11,8 @@ Features
   - Allow empty ``in-path-class`` which then does not set the in-path class.
   - Allow configuration of ``item-wrapper`` element on which ``in-path-class`` or ``current-class`` are set. Defaults to ``li``.
   - Set current class also on parent item-wrapper if not set.
+  - Fix case where current class was not set when not present on load.
+  - Set current class also when DOM elements are inserted or removed.
 - Add ``utils.getCSSValue`` for retrieving CSS property values for DOM nodes.
 - Add configurable scrolling behavior to pat-inject.
 - Add ``webpack-visualizer-plugin`` for analyzation of generated bundles.